Fix inventory slot remapping

+++ src/protocolsupport/protocol/packet/middle/clientbound/play/MiddleInventoryHorseOpen.java
@@ -16,7 +16,7 @@ public abstract class MiddleInventoryHorseOpen extends ClientBoundMiddlePacket {
 		super(connection);
 	}
 
-	protected int windowId;
+	protected byte windowId;
 	protected int slots;
 	protected int entityId;
